♻️ Move useShareLivePosition
